Add pkg and cmd folder and reorganise backend code
This commit adds a top level pkg and cmd directory, and moves the
backend code into that directory. This is the first commit reorganising
the code under a common pkg and cmd directories in preparation for
splitting the code into microservices.
