+++ pkg/backend/database/database.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/openclarity/vmclarity/backend/pkg/database/gorm"
-	"github.com/openclarity/vmclarity/backend/pkg/database/types"
+	"github.com/openclarity/vmclarity/pkg/backend/database/gorm"
+	"github.com/openclarity/vmclarity/pkg/backend/database/types"
 )
 
 type DBDriver func(config types.DBConfig) (types.Database, error)
