+++ pkg/backend/database/gorm/finding.go
@@ -24,8 +24,8 @@ import (
 	"gorm.io/gorm"
 
 	"github.com/openclarity/vmclarity/api/models"
-	"github.com/openclarity/vmclarity/backend/pkg/common"
-	"github.com/openclarity/vmclarity/backend/pkg/database/types"
+	"github.com/openclarity/vmclarity/pkg/backend/common"
+	"github.com/openclarity/vmclarity/pkg/backend/database/types"
 )
 
 type Finding struct {
